Add config for vimwiki plugin and a pandoc command
Changing some default settings of vimwiki, namely the location of the
default index file, and the format to use (markdown).

Setting a custom command `Pan` which runs `pandoc` on the current file,
saves it as html (with the same root file name), and opens it in
the browser.

Also changed the `EditRC` command to show my vimrc in a split instead of
a new tab, as Windows does weird things when it opens a new tab.